Split conftest into multiple files
Splitting conftest makes it easier to group related fixtures, while not having the single file grow to be giant.
This will be particularly useful for generating version numbers.

This doesn't seem to currently work with python2.
